+++ src/main/java/me/kall/narutoloading/mixin/EarlySetup.java
@@ -1,7 +1,9 @@
 package me.kall.narutoloading.mixin;
 
 import com.google.common.collect.Lists;
+import me.kall.narutoloading.data.FFmpeg;
 import me.kall.narutoloading.data.NarutoConfig;
+import me.kall.narutoloading.data.SourceRoller;
 import net.minecraftforge.fml.loading.FMLLoader;
 import org.objectweb.asm.tree.ClassNode;
 import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
@@ -15,6 +17,8 @@ public class EarlySetup implements IMixinConfigPlugin {
     public void onLoad(String s) {
         if (FMLLoader.getDist().isClient()) {
             NarutoConfig.init();
+            SourceRoller.init();
+            FFmpeg.init();
         }
     }